Drop requirement for certifi in PROD. Drop explicit google-cloud-bigquery pinning.
