+++ scripts/common_test.py
@@ -1352,6 +1352,11 @@ class UrlRetrieveTests(CommonTests):
         self.swap_curl_failure = self.swap(
             subprocess, 'Popen', mock_failing_curl_popen)
 
+    def test_url_open(self) -> None:
+        response = common.url_open('http://www.google.com')
+        self.assertEqual(response.getcode(), 200)
+        self.assertEqual(response.url, 'http://www.google.com')
+
     def test_url_retrieve_tries_curl_at_outset(self) -> None:
         with tempfile.TemporaryDirectory() as tempdir:
             output_path = os.path.join(tempdir, 'output')
